requirements.txt
pandas
transformers
datasets
sentencepiece
accelerate
pyyaml
nltk
openai
tqdm
google-genai
xlsxwriter
# torch

# Testing dependencies
pytest>=7.0
pytest-mock>=3.10
pytest-cov>=4.0
pytest-timeout>=2.1